First version, just need to be tested with a real image

+++ build.sh
@@ -1,30 +1,56 @@
 #!/bin/bash
 
-if [ -f /tmp/fos-usb.img ]; then
-    rm -f /tmp/fos-usb.img
+if [ "$EUID" -ne 0 ]; then
+    echo "Please run as root"
+    exit
 fi
 
-echo Make a blank 128MB disk image
-dd if=/dev/zero of=/tmp/fos-usb.img bs=1M count=128
 mkdir -p cache
+if [ -f cache/fos-usb.img ]; then
+    rm -f cache/fos-usb.img
+fi
 
-echo Download the FOG kernels and inits
-wget -N -P cache/ https://github.com/FOGProject/fos/releases/download/20230331/bzImage
-wget -N -P cache/ https://github.com/FOGProject/fos/releases/download/20230331/init.xz
+if [ ! -f cache/bzImage ]; then
+    wget -N -P cache/ https://github.com/FOGProject/fos/releases/download/20230331/bzImage
+fi
+if [ ! -f cache/init.xz ]; then
+    wget -N -P cache/ https://github.com/FOGProject/fos/releases/download/20230331/init.xz
+fi
 
+
+echo Create custom init
+unxz --verbose -kc cache/init.xz > /tmp/fos-init
+LOOPINIT=$(losetup --show -f /tmp/fos-init)
+MOUNTINIT=$(mktemp -d)
+mount $LOOPINIT $MOUNTINIT
+cp -prv patch/* $MOUNTINIT/
+umount $MOUNTINIT
+rmdir $MOUNTINIT
+losetup -d $LOOPINIT
+xz --check=crc32 --verbose -kc /tmp/fos-init > cache/custom-init.xz
+
+echo Make a blank disk image
+dd if=/dev/zero of=cache/fos-usb.img bs=1M count=256
  
 echo Make the partition table, partition and set it bootable.
-parted --script /tmp/fos-usb.img mklabel msdos mkpart primary fat32 1 128 set 1 boot on
- 
-echo Map the partitions from the image file
-kpartx -a -s /tmp/fos-usb.img
-LOOPDEV=$(losetup -a | grep "/tmp/fos-usb.img" | grep -o "loop[0-9]*")
+parted --script cache/fos-usb.img mklabel msdos mkpart primary fat32 1 128 set 1 boot on mkpart primary ext2 129 256
 
+echo Map the partitions from the image file
+kpartx -a -s cache/fos-usb.img
+LOOPDEV=$(losetup -a | grep "cache/fos-usb.img" | grep -o "loop[0-9]*")
+
+if [ -z "${LOOPDEV}" ]; then
+    echo "Failed to setup loop!"
+    exit
+fi
  
-echo Make an vfat filesystem on the first partition.
+echo Make filesystems
 mkfs -t vfat -n GRUB /dev/mapper/${LOOPDEV}p1
+mkfs -t ext2 -L IMAGES /dev/mapper/${LOOPDEV}p2
+# fstab: LABEL=IMAGES
  
 echo Mount the filesystem via loopback
+MOUNT=$(mktemp -d)
 mount /dev/mapper/${LOOPDEV}p1 $MOUNT
 
 echo Install GRUB
@@ -34,30 +60,27 @@ grub-install --removable --no-floppy --boot-directory=$MOUNT/boot --target=i386-
 
 echo Copy boot files
 cp cache/bzImage $MOUNT/boot/
-cp cache/init.xz $MOUNT/boot/
+cp cache/custom-init.xz $MOUNT/boot/
 
 echo Create the grub configuration file
 cat > $MOUNT/boot/grub/grub.cfg << 'EOF'
 
-set myfogip=http://192.168.1.100
-set myimage=/boot/bzImage
-set myinits=/boot/init.xz
 set timeout=-1
 insmod all_video
 
 menuentry "FOG Deploy from USB" {
- echo loading the kernel
- linux  $myimage loglevel=4 initrd=init.xz root=/dev/ram0 rw ramdisk_size=275000 keymap= web=$myfogip/fog/ boottype=usb consoleblank=0 rootfstype=ext4
- echo loading the virtual hard drive
- initrd $myinits
+ echo loading kernel
+ linux /boot/bzImage loglevel=4 initrd=custom-init.xz root=/dev/ram0 rw ramdisk_size=275000 keymap= boottype=usb consoleblank=0 rootfstype=ext4
+ echo loading ram disk
+ initrd /boot/custom-init.xz
  echo booting kernel...
 }
 
 menuentry "[DEBUG] FOG Deploy from USB" {
- echo loading the kernel
- linux  $myimage loglevel=7 initrd=init.xz root=/dev/ram0 rw ramdisk_size=275000 keymap= web=$myfogip/fog/ boottype=usb consoleblank=0 rootfstype=ext4 isdebug=yes
- echo loading the virtual hard drive
- initrd $myinits
+ echo loading kernel
+ linux /boot/bzImage loglevel=7 initrd=custom-init.xz root=/dev/ram0 rw ramdisk_size=275000 keymap= boottype=usb consoleblank=0 rootfstype=ext4 console=tty0 console=ttyS0,115200n8 isdebug=yes
+ echo loading ram disk
+ initrd /boot/custom-init.xz
  echo booting kernel...
 }
 
@@ -68,4 +91,4 @@ umount $MOUNT
 rmdir $MOUNT
  
 echo Unmap the image
-kpartx -d /tmp/fos-usb.img
+kpartx -d cache/fos-usb.img
